Add files via upload
This is a simulation where you can get to know whether a vehicle accident is lethal or safe. You can assume the square sprite as a vehicle and the long rectangular sprite as the wall. If you have finished the first outcome, press the sprite to make the sprite go to its original position. And press the SPACE to run the simulation.